+++ lib/app.js
@@ -1,6 +1,6 @@
 var Git            = require('./gitmech')
-  , Tools          = require('./tools')
-  , Components     = require('./components');
+  , Components     = require('./components')
+  , gravatar       = require('gravatar');
 
 
 function requireAuthentication(req, res, next) {
@@ -10,6 +10,33 @@ function requireAuthentication(req, res, next) {
     next();
   }
 }
+
+function initJingoResponse(req, res, next) {
+  res.locals({
+    get user() {
+      return req.user;
+    },
+    isAnonymous: function () {
+      return !req.user;
+    },
+    canSearch: function () {
+      return !!req.user || app.locals.authorization.anonRead;
+    },
+    gravatar: function(email) {
+      return gravatar;
+    },
+    hasSidebar: Components.hasSidebar,
+    hasFooter: Components.hasFooter,
+    hasCustomStyle: Components.hasCustomStyle,
+    hasCustomScript: Components.hasCustomScript,
+    hasFeature: function(feature) {
+      return !!app.locals.features[feature];
+    }
+  });
+
+  next();
+};
+
 
 function setupRemotePush(app) {
   if (app.locals.remote != "") {
@@ -46,30 +73,30 @@ function customStuff(req, res, next) {
   next();
 }
 
-  
+
 module.exports.setup = function(app, config) {
   var gitrepopath = config.jingo_gitrepo_path || process.cwd() + '/data';
   var giktsubdir = config.jingo_gitrepo_subdir || '';
-  
+
   try {
     Git.setup(gitrepopath, giktsubdir);
   } catch(e) {
     console.log(e.message);
     process.exit(-1);
   }
-  
+
   app.locals.Git = Git;
   app.locals.Components = Components;
   app.locals.appTitle = config.jingo_app_title || "Jingo";
   app.locals.features = config.jingo_features || {};
-  
+
   // For backward compatibility with version < 0.5, we set markitup as the
   // default rich editor. For new installations, the default is codemirror
   if ("codemirror" in app.locals.features &&
       "markitup" in app.locals.features) {
     app.locals.features.markitup = true;
   }
-  
+
   var refspec = config.jingo_appremote ? config.jingo_appremote.split(/\s+/) : "";
   if (!refspec) {
     app.locals.remote = "";
@@ -78,46 +105,27 @@ module.exports.setup = function(app, config) {
     app.locals.remote = refspec[0].trim();
     app.locals.branch = refspec[1] ? refspec[1].trim() : "master";
   }
-  
+
   app.locals.pushInterval = config.jingo_push_interval ? parseInt(config.jingo_push_interval, 10) * 1000 : 30000;
   app.locals.coalesce = function(value, def) {
     return typeof value === 'undefined' ? def : value;
   };
-  
+
   Components.init(Git);
-  
+
+  app.use(initJingoResponse);
   app.use(customStuff);
-  
+
   app.all("/pages/*", requireAuthentication);
-  
+
   var anonread = typeof config.anonRead === 'undefined' ? true : config.anonRead;
   if (!anonread) {
     app.all("/wiki/*", requireAuthentication);
     app.all("/search", requireAuthentication);
   }
-  
-  var routes = require("./routes");
-  app.get    ("/",                      routes.index);
-  app.get    ("/wiki",                  routes.pageList);
-  app.get    ("/wiki/:page",            routes.pageShow);
-  app.get    ("/wiki/:page/history",    routes.pageHistory);
-  app.get    ("/wiki/:page/:version",   routes.pageShow);
-  app.get    ("/wiki/:page/compare/:revisions", routes.pageCompare);
-  
-  app.get    ("/search",                routes.search);
-  
-  app.get    ("/pages/new",             routes.pageNew);
-  app.get    ("/pages/new/:page",       routes.pageNew);
-  app.post   ("/pages",                 routes.pageCreate);
-  
-  app.get    ("/pages/:page/edit",      routes.pageEdit);
-  app.put    ("/pages/:page",           routes.pageUpdate);
-  app.del    ("/pages/:page",           routes.pageDestroy);
-  
-  app.post   ("/misc/preview",          routes.miscPreview);
-  app.get    ("/misc/syntax-reference", routes.miscSyntaxReference);
-  app.get    ("/misc/existence",        routes.miscExistence);
-  
+
+  require("./routes").setup(app);
+
   setupRemotePush(app);
-  
+
 };
